+++ NEWS.md
@@ -114,6 +114,7 @@
 - New methods `$select_seq()` and `$with_columns_seq()` for `DataFrame` and
   `LazyFrame` (#1003).
 - New method `$clear()` for `DataFrame`, `LazyFrame`, and `Series` (#1004).
+- New method `$struct$unnest()` for `Series` (#1010).
 
 ### Bug fixes